Added camera API
