+++ vision/__init__.py
@@ -1,0 +1,6 @@
+from .camera import Camera
+from .main import VisionModule
+
+__all__ = [
+    "Camera", "VisionModule"
+]
