+++ vision/camera.py
@@ -1,0 +1,64 @@
+"""Camera interface module
+
+Usage
+-----
+camera = Camera()
+img = camera.capture()
+
+# do things with img
+camera.save()  # saves current image
+
+camera.close()
+"""
+
+import time
+
+import cv2
+from picamera import PiCamera
+from picamera.array import PiRGBArray
+
+
+class Camera:
+    """PiCamera interface"""
+
+    def __init__(self):
+        self.camera = PiCamera()
+        self.camera.resolution = (640, 480)
+        self.camera.awb_mode = 'fluorescent'
+        self.capture_raw = PiRGBArray(self.camera)
+
+        self.frame_id = 0
+        self.fps = 0
+        self.start_time = time.time()
+
+    def capture(self):
+        """Capture image
+
+        Returns
+        -------
+        np.array
+            reference to image array; NOT UNIQUE PER CAPTURE.
+        """
+
+        self.camera.capture(
+            self.capture_raw, format='bgr', use_video_mode=True)
+
+        self.frame_id += 1
+        self.fps = (time.time() - self.start_time) / self.frame_id
+
+        return self.capture_raw.array
+
+    def save(self):
+        """Save current frame"""
+
+        cv2.imwrite("{}.jpg".format(self.frame_id), self.capture_raw.array)
+
+    def close(self):
+        """Close camera"""
+
+        self.camera.close()
+
+    def __del__(self):
+        """Destructor method to ensure camera closing"""
+
+        self.close()
